rename file & fix account_manager
in folder service, rename crudUser.jsx to userService
